refactor: dissolve @kolu/port-scan into osfacts-client + padi policy
The OS volatility that justified a package lives in osfacts; the TypeScript
face is osfacts-client (no @kolu scope, zero npm runtime deps — drishti is
the next consumer). Kolu policy (addressBind, U-row blindness,
scanSubtreePorts, KOLU_OSFACTS_BIN) moves to padi; PortInfo/foldPorts home
to terminal-vocab. Delete packages/port-scan with no shim.
