+++ docs/atlas/src/content/atlas/port-forwarding.mdx
@@ -268,19 +268,18 @@ data arrives on the terminal's metadata like everything else). The flows:
 
 ## Detection: ports, attributed to the terminal
 
-The scanner is its own package — **`@kolu/port-scan`**
-(`packages/port-scan/`, with the darwin C helper in its `native/`) — driven by a
-**padi sensor**, and not a kaval module. Two separate calls, and both matter.
+The scanner is **padi's port scan** (`packages/padi/src/portScan.ts` +
+`portSampler.ts`), driven by a **padi sensor**, not a kaval module. The OS
+volatility that once justified `@kolu/port-scan` as its own package now lives
+in [osfacts](./os-facts-tool.html); the TypeScript face is
+`osfacts-client` (`osfacts/client-ts/`), and kolu policy (bind classification,
+U-row blindness, fold) sits beside the sampler in padi. Two separate calls,
+and both matter.
 
-*Why a package*: the OS is the volatility. How a kernel names the processes
-holding listening sockets has already changed three times on darwin in this
-repo alone, and linux answers by a different mechanism in a different byte
-order. That is the [electricity](electricity) bar — a real axis of change
-behind a stable receptacle — so the reader lives outside padi with the arrow
-pointing out, and its only input is a list of root pids. Test ③ (a *second*
-consumer) is met from both sides now: vazhi shipped in PRT0, and
-[osfacts](./os-facts-tool.html) — whose OSF2 phase adopts it as this
-scanner's engine — is built.
+*Why not a package any more*: the OS axis of change is osfacts's job. padi
+owns the cadence and the policy; the client is the tool's own leaf (and
+drishti's next import). Test ③ (a *second* consumer) is met by
+`osfacts-client` + the imminent drishti adoption.
 
 *Why padi drives it*: the repo's own taxonomy draws the line. Facts that require the
 PTY itself live in kaval (foreground needs `tcgetpgrp` on the PTY fd, cwd is
@@ -299,14 +298,14 @@ which is one more reason the scanner lives there.</Footnote> builds the
 pid→ppid table once, partitions it into per-shell subtrees, and joins
 listening TCP sockets against each subtree:
 
-- **linux** — pid table from `/proc/*/stat`; LISTEN rows (with their **bind address**) from `/proc/net/tcp` + `tcp6`, joined to pids via `/proc/<pid>/fd` socket-inode readlinks (the same technique `socketHolder.ts` uses — RE-DERIVED, not shared: one leaf both plug into, with the fd-failure policy injected, is a standing extraction), walking **only the union of terminal subtrees**. Expect and tolerate per-pid `EACCES`/`ENOENT` outside the subtrees — but `EACCES` on a *requested* subtree surfaces as a scan error, never an empty port list. Do not spawn `ss` (measured ~7× slower, and it hides other-users' pid attribution anyway).
-- **darwin** — a **libproc helper kolu builds** (`packages/port-scan/native/portScanDarwin.c`, compiled by the nix derivation beside it, its path baked onto padi as `KOLU_PORT_SCAN_HELPER` with no `PATH` fallback and no system fallback). ONE invocation prints BOTH tables in a versioned TSV — `V\t1`, then `P\tpid\tppid\tname` and `L\tpid\tport\thex-bind-address` (tab-separated) — and refuses a version it does not know, because a half-understood shape would parse to zero listeners and read as “nothing here is serving”. It emits **raw address bytes, not a wildcard boolean**, so `isAnyAddress` stays the single judge for both platforms; that hex is in NETWORK order where `/proc`’s is host-order-per-32-bit-word, which is why there are two decoders and exactly one judge. Non-root sees own-uid pids only — sufficient by construction, since a terminal’s subtree is padi’s own uid.<Footnote>This replaced `ps` + `lsof` (~143 ms of subprocess per
-pass), which had replaced `netstat` — disqualified because macOS 27's
-netstat intermittently loses its internet tables while reporting success
-(verified across three sampling windows on one boot; cause unknown), and its
-fixed-width columns truncate the bind address reachability is judged from.
-"Answers no ports, successfully" is the worst failure mode for a feature
-whose whole job is that question.</Footnote>
+- **both platforms** — one spawn of the baked **osfacts** binary
+  (`KOLU_OSFACTS_BIN`, via `osfacts-client`) with `--roots` over the terminal
+  root pids. Versioned TSV (`V 1`, then `P` / `L` / `U` rows); raw network-order
+  bind-address hex; padi's `addressBind` is the single scope/family judge.
+  A `U` on a *requested* root is a blind scan (never an empty port list); a `U`
+  on a descendant is skipped (the sudo lesson). Non-root sees own-uid pids only
+  on darwin — sufficient by construction, since a terminal's subtree is padi's
+  own uid.
 
 
 ### The successor: osfacts
@@ -364,7 +363,7 @@ and spells the wildcard bind `*`. One more, found live: `/proc/<pid>/stat`'s
 `/proc/<pid>/cmdline`'s `argv[0]` instead.</Footnote>
 
 Each terminal's sample — `PortInfo { port, name, scope, family }` in
-`packages/port-scan/src/ports.ts` (re-exported by `terminal-vocab`), beside `ForegroundSchema:145` — is
+`@kolu/terminal-vocab` (`ports.ts`), beside `ForegroundSchema:145` — is
 **deduped by structural equality** before it touches the snapshot, so an
 unchanged scan emits nothing: the churn guard that keeps a 5-second ticker
 from thrashing SolidJS. From there every layer exists: the sensor set folds
@@ -415,7 +414,7 @@ answering, the property the per-forward connection exists to give.
 
 ### PRT1 — detection + Ports section (shipped: <PrLink pr={1982} />)
 
-Scanner (`@kolu/port-scan`), the `ports` schema field, the sensor fold,
+Scanner (padi `portScan` + `osfacts-client`), the `ports` schema field, the sensor fold,
 the Inspector section, direct-open chips, the tip. What remains load-bearing:
 the pass holds **no OS state between ticks** (parse, join, emit the
 structural `PortInfo`, drop everything else); it **repartitions from the
