+++ osfacts/client-ts/package.json
@@ -1,26 +1,20 @@
 {
-  "name": "@kolu/port-scan",
+  "name": "osfacts-client",
   "version": "0.1.0",
   "private": true,
-  "license": "AGPL-3.0-or-later",
+  "license": "MIT OR Apache-2.0",
   "type": "module",
   "sideEffects": false,
   "main": "./src/index.ts",
   "types": "./src/index.ts",
   "exports": {
-    ".": "./src/index.ts",
-    "./ports": "./src/ports.ts"
+    ".": "./src/index.ts"
   },
   "scripts": {
     "typecheck": "tsc --noEmit",
     "test:unit": "vitest run"
   },
-  "dependencies": {
-    "@kolu/log": "workspace:*",
-    "zod": "^4.3.6"
-  },
   "devDependencies": {
-    "@kolu/daemon-test-gate": "workspace:*",
     "@types/node": "^22.0.0",
     "typescript": "^7.0.2",
     "vitest": "^4.1.0"
