+++ packages/padi/package.json
@@ -27,7 +27,6 @@
     "test:unit": "vitest run"
   },
   "dependencies": {
-    "@kolu/port-scan": "workspace:*",
     "@kolu/serve-dir": "workspace:*",
     "@kolu/surface": "workspace:*",
     "@kolu/surface-daemon": "workspace:*",
@@ -51,6 +50,7 @@
     "kolu-pty": "workspace:*",
     "kolu-transcript-core": "workspace:*",
     "kolu-transcript-html": "workspace:*",
+    "osfacts-client": "workspace:*",
     "pino": "^10.3.1",
     "pino-pretty": "^13.1.3",
     "pino-roll": "^3.1.0",
