+++ packages/server/package.json
@@ -17,7 +17,6 @@
     "@kolu/log": "workspace:*",
     "@kolu/padi": "workspace:*",
     "@kolu/port-forward": "workspace:*",
-    "@kolu/port-scan": "workspace:*",
     "@kolu/serve-dir": "workspace:*",
     "@kolu/surface": "workspace:*",
     "@kolu/surface-app": "workspace:*",
