+++ packages/surface-daemon-supervisor/src/socketHolder.ts
@@ -44,7 +44,7 @@ const execFileP = promisify(execFile);
 /** Hard ceiling on the darwin `lsof` subprocess — a wedged `lsof` (contended mount,
  *  slow box) must be killed and rejected, never left to hang the serving event loop.
  *
- *  The same number as `@kolu/port-scan`'s `PORT_SCAN_COMMAND_TIMEOUT_MS` (`scan.ts`), which
+ *  The same number as padi's `PORT_SCAN_COMMAND_TIMEOUT_MS` (`portScan.ts`), which
  *  runs the same binary for its own question; the two are cross-named because the
  *  socket↔pid reading is re-derived in both places rather than shared. */
 const LSOF_TIMEOUT_MS = 5_000;
@@ -52,7 +52,7 @@ const LSOF_TIMEOUT_MS = 5_000;
 /** macOS's OWN `lsof`, by ABSOLUTE path. kolu's macOS users run nix, so `lsof` on
  *  `PATH` may be a different build than the system one this parse was verified
  *  against — resolving it through `PATH` would make a boot-path check's correctness
- *  depend on the user's profile. Same call, same reason, as `@kolu/port-scan`'s
+ *  depend on the user's profile. Same call, same reason, as padi port-scan's
  *  `DARWIN_LSOF`. */
 const DARWIN_LSOF = "/usr/sbin/lsof";
